# optimize add by ~30%