old: 

```
test u256_add      ... bench:     135,355 ns/iter (+/- 30,832)
test u512_add      ... bench:     213,261 ns/iter (+/- 44,027)
```

new:

```
test u256_add      ... bench:     107,183 ns/iter (+/- 14,998)
test u512_add      ... bench:     148,255 ns/iter (+/- 25,584)
```